module_logging_notification/requirement.txt
google-cloud-pubsub 
google-cloud-logging
google-cloud-secret-manager 
requests 
# HTTP (Telegram Bot API)
